+++ src/main/java/org/jolokia/docker/maven/access/chunked/ChunkedResponseHandler.java
@@ -1,7 +1,0 @@
-package org.jolokia.docker.maven.access.chunked;
-
-import org.jolokia.docker.maven.access.DockerAccessException;
-
-public interface ChunkedResponseHandler<T> {
-    void process(T toProcess) throws DockerAccessException;
-}